requirements.txt
discord.py==2.3.2
aiohttp
requests
python-dotenv
typing_extensions
async-timeout
flask
